extract the Profile class to a new file
